feat: back btn handling for main screens
